remove unnecessary directories by moving stuff into lib/src
