+++ packages/patrol_cli/lib/src/devices.dart
@@ -5,7 +5,6 @@ import 'package:meta/meta.dart';
 import 'package:patrol_cli/src/base/exceptions.dart';
 import 'package:patrol_cli/src/base/logger.dart';
 import 'package:patrol_cli/src/base/process.dart';
-import 'package:patrol_cli/src/features/devices/device.dart';
 import 'package:process/process.dart';
 
 class DeviceFinder {
@@ -141,6 +140,71 @@ class DeviceFinder {
       throwToolInterrupted('`flutter devices` was interrupted');
     } else {
       return output;
+    }
+  }
+}
+
+class Device {
+  const Device({
+    required this.name,
+    required this.id,
+    required this.targetPlatform,
+    required this.real,
+  });
+
+  final String name;
+  final String id;
+  final TargetPlatform targetPlatform;
+  final bool real;
+
+  /// Returns the name that Patrol is usually interested with.
+  ///
+  /// On Android, this is the ID of the device, e.g "emulator-5554".
+  ///
+  /// On iOS, this is the name of the device, e.g "iPhone 13" or
+  /// "Barteks-iPhone".
+  String get resolvedName {
+    switch (targetPlatform) {
+      case TargetPlatform.android:
+        return id;
+      case TargetPlatform.iOS:
+        return name;
+    }
+  }
+
+  String get description {
+    switch (targetPlatform) {
+      case TargetPlatform.android:
+        if (platformDescription.isEmpty) {
+          return id;
+        } else {
+          return '$platformDescription $id';
+        }
+      case TargetPlatform.iOS:
+        return '$platformDescription $name';
+    }
+  }
+
+  String get platformDescription {
+    switch (targetPlatform) {
+      case TargetPlatform.android:
+        return real ? 'device' : '';
+      case TargetPlatform.iOS:
+        return real ? 'device' : 'simulator';
+    }
+  }
+}
+
+enum TargetPlatform { iOS, android }
+
+extension TargetPlatformX on TargetPlatform {
+  static TargetPlatform fromString(String platform) {
+    if (platform == 'ios') {
+      return TargetPlatform.iOS;
+    } else if (platform.startsWith('android-')) {
+      return TargetPlatform.android;
+    } else {
+      throw Exception('Unsupported platform $platform');
     }
   }
 }
