+++ packages/patrol_cli/lib/src/features/test/native_test_runner.dart
@@ -1,5 +1,5 @@
 import 'package:dispose_scope/dispose_scope.dart';
-import 'package:patrol_cli/src/features/devices/device.dart';
+import 'package:patrol_cli/src/devices.dart';
 import 'package:patrol_cli/src/test_runner.dart';
 
 class NativeTestRunner extends TestRunner implements Disposable {
