+++ packages/patrol_cli/lib/src/pubspec_reader.dart
