+++ packages/patrol_cli/test/features/run_commons/device_finder_test.dart
@@ -1,6 +1,6 @@
 import 'package:dispose_scope/dispose_scope.dart';
 import 'package:patrol_cli/src/base/exceptions.dart';
-import 'package:patrol_cli/src/features/devices/device_finder.dart';
+import 'package:patrol_cli/src/devices.dart';
 import 'package:test/test.dart';
 
 import '../../fixtures.dart';
